Spring Security - Autenticacion añadida (Faltan detalles con las condicionales)

+++ src/main/java/com/ambystudio/forohub/domain/topicos/DTOListarTopico.java
@@ -1,6 +1,6 @@
 package com.ambystudio.forohub.domain.topicos;
 
-public record DTOListarTopico(String titulo, String mensaje, String status, Integer autor, Integer curso) {
+public record DTOListarTopico(String titulo, String mensaje, String status, Long autor, Long curso) {
 
     public DTOListarTopico(Topico topico){
         this(topico.getTitulo(), topico.getMensaje(), topico.getStatus(), topico.getAutor(), topico.getCurso());
